adding some unit tests also upgraded libraries, had to reinterpret_cast a string to const char * as it was unsigned.

+++ src/opengl/glengine/BlendModeGeneric.hpp
@@ -29,9 +29,10 @@ public:
     {
       return 0;
     }
-    return selection_values[static_cast<int>(std::clamp(
+    return selection_values[static_cast<std::size_t>(std::clamp(
       selection,
-      int{},      static_cast<int>(std::ranges::ssize(selection_values) - 1)))];
+      int{},
+      static_cast<int>(std::ranges::ssize(selection_values) - 1)))];
   }
 
   auto &operator=(int new_selection) noexcept
